reorganizing pynumero slightly
