+++ pyomo/contrib/pynumero/linalg/tests/test_ma57.py
@@ -7,18 +7,12 @@
 #  rights in this software.
 #  This software is distributed under the 3-clause BSD License.
 #  ___________________________________________________________________________
-import sys
-import os
 import ctypes
 import pyutilib.th as unittest
-
 from pyomo.contrib.pynumero.dependencies import numpy as np, numpy_available
 if not numpy_available:
     raise unittest.SkipTest('pynumero MA27 tests require numpy')
-
-import numpy.ctypeslib as npct
-
-from pyomo.contrib.pynumero.extensions.ma57_interface import *
+from pyomo.contrib.pynumero.linalg.ma57 import *
 
 
 @unittest.skipIf(not MA57Interface.available(), reason='MA57 not available')
